Update README.md
I have built a simple obesity prediction ml model using a simple dataset. To access the model in a webpage. I've built the Flask framework with flasgger and SwaggerUI API. Finally, I've built a simple docker image and dockerized it to access it across different environments.

Docker
------

What are containers?
For standardization of Environments.
Containerization is a way to package all the necessary dependencies and configurations.
Portable artifact - easily share and move this package to any environment.
Makes the development and deployment easy to sync.
Layers of images = Docker Image
If we run this docker image, the Docker will create a container to install all environments and dependencies to run the application.
Virtual Machines will interact with the application layer and OS layer, whereas, the Docker Image will interact with only the application layer and not the OS layer.
The size of the Dockers will be small compared to Virtual Machines.
Dockers are fast compared to VM.
We can install any VM on any OS.
Docker Images of a specific OS are not compatible with another OS.


0.0.0.0 port number can access the local address and local IP address also.

Docker commands
---------------
FROM - pull a specific docker image from the docker repository.
COPY - all the files from the current directory should be copied to the docker image
RUN - install all the libraries from requirements.txt
CMD - run the docker file

docker build -t dockerimagename . - command to build a docker image with a specific name.
docker run -p 5000:5000 dockerimagename - run docker image with mapping the port of the host to that of the container.
docker ps - view the running containers.
docker stop containerid - stop the docker container
docker push dockerimagename:latest - push the docker image.
docker pull dockerimagename:latest - pull the docker image.